fixes #770 (in a better way)
I've modified the solution to #770 by overriding the dictionary class of the dictionary of variables passed to the eval() function. This allows us to trace which globals a particular global depends on. This was particularly important because of the following issue:

Imagine a global that was defined as: 7 if condition else a_list_of_values

In the previous code, the global would be found to depend on a_list_of_values regardless of the state of condition. In the new code, it determines dependencies based on what the eval() function requests access to, and so the dependencies do depend on the state of condition as one would expect.

The only "strange" behaviour now is if you set a global to: list_1 if za_list_of_conditions else list_2
In this case, the global will be added to a zip group with 'za_list_of_conditions' (note that "if za_list_of_conditions" always evaluates to True if it is a non empty list) rather than the more obvious option of adding it to a zip group with 'list_1'. Whether the global is added to a zip group with 'list_1' or 'za_list_of_conditions' depends on the global names (the zip group picked is the last dependency in alphabetical order). You can of course, always override the runmanager defaults yourself if you ever encounter an instance like this!