Fixed bug where experiment folder name would be incorrect
If the labscript file name (excluding extension) ended with the letters 'p' or 'y', these letters would be missing from the folder name automatically generated by runmanager for storing experiments. This is because we were using strip('.py') which removes the *list* of characters specified from the end of a string until a character not in the list was found.
For instance the following statements are True:
'exp.py'.strip('.py') == 'ex'
'expyy.ypypp.py'.strip('.py') == 'ex'