Qt: Resolves bitbucket issue #3. The top three tabs are now not movable.
This was a little tricky to do because Qt does not allow for setting individual
tabs movable or not. Furthermore, the relevant methods one has to override on
the QTabBar contain a lot of calls to private methods, which makes re-implementing
their behaviour more work than is worth it.

So instead, I did a bit of a hack - the subclassed TabBar keeps track of whether
its individual tabs are supposed to be movable or not, and it turns the global
setting on and off on mouseclick and mouserelease events, ensuring that when
you click an unmovable tab, you will find that it is unmovable and when you click
a movable tab, that it is movable.

It seems to work.