pstarkey/bugfix: Fixed bug where compilation of single shot (aka connection table recompilation) using a globals file that contained globals of lists/arrays resulted in a confusing error.
Previously, the error listed the names of every global group provided that group contains more than one global. This is because we changed the return result of `evaluate_globals` without updating this code.

Now we return the names of globals where the "expansion" column is set. I think this might ignore automatic zip groups, but that probably makes sense as you really only care about the globals that you need to edit, which would still be listed as 'outer' since we don't run the automatic zip group checking code in the API (in fact that code (`guess_expansion_modes`) only exists in the GUI part of runmanager, which means the API/GUI are not actually equivalent when compiling a sequence of shots. Not relevant here....but I'll log an issue about it anyway!)