bugfix: Remove use of NetworkOnlyLock, which no longer exists in h5_lock. This was an accidental break in backward compatibility.
NetworkOnlyLocks worked by acquiring a zlock lock once before many openings
and closings og a h5 file, and those openings and closings used regular in-process
threading.RLocks. This reduced network traffic.

The complexity of reimplementing NetworkOnlyLocks is not worth the slight
performance improvement in my opinion. It would be easier to refactor how
runmanager opens globals to pass around an already opened file instead. It is
always possible to refactor code in this way, so that should be the preferred
approach if we encounter performance issues with repeated file openings in
the future.