word_wrap: Word-wrapping in treeviews and item views, adapted from David Meyer's implementation.
Swapped the globals treeview for a tableview. This involved replacing the custom
treeview with a mixin class, so that it could be reused for both the new tableview
and the existing treeviews.

This was because TableViews have automatic vertical resizing of rows to contents,
whereas treeviews do not.

Added a custom editor for globals that also has word wrapping and expands
vertically when wrapping occurs. The editor is a QTextEdit instead of a QPlainTextEdit
because QPlainTextEdit has worse resizing semantics, preferring to return sizes in
number of lines instead of pixels, and always having one additional blank line
of space reported in its size. QTextEdit is simple enough to use instead, we
just have to explicitly say we want to set and get plain text only.

Disabled resizing of all columns that were still possible to resize. All columns
are not set to their minimum width except for one column in each view that is
set to stretch, which in the case of the globals view is the value column.

Most of the diff is noise of variable renamings due to changing the treeview
for a tableview, and boilerplate changes due to using a mixin. The changes of
substance are the new TableView class definition, the new Editor class defintion,
and some changes to the ItemDelegate (which I also simplified to not use
font metrics in its height calculation). The other meaningful changes are setting
a stretch section for the views and setting all other sections to minimum column
width.