runmanager-mono-font: Use consolas as the small monospaced font on Windows rather than Ubuntu.
Qt's font rendering on windows'seems to be more variable, and at least on Windows 7,
the Ubuntu font at smaller sizes looks absolutely horrendous. On Windows 10
it is a little better, but still with Qt and whatever settings I have, there
doesn't seem to be any sub-pixel smoothing.

Consolas is wider at the same height and so more forgiving of bad font rendering.
It should be present on all Windows 7 and Windows 10 computers.